backend/requirements.txt
flask
flask_cors
gunicorn
requests
beautifulsoup4
selenium
webdriver-manager
python-dotenv
redis
httpx
anyio